Replace matrix diagonals

Sets a band of matrix diagonals to any given value

bandReplace(mat, lower, upper, value = NA, silent = TRUE)

Arguments

mat
A Matrix

lower
Lower diagonal to be included in the band (should be \(\le 0\))

upper
Upper diagonal to be included in the band (should be \(\ge 0\))

value
A single value to replace all values in the selected band (default = NA)

silent
Operate in silence, only (some) warnings will be shown (default = TRUE)

Value

A matrix in which the values in the selected diagonals have been replaced

See also

Examples

# Create a 10 by 10 matrix
library(Matrix)
m <- Matrix(rnorm(10),10,10)

bandReplace(m,-1,1,0)   # Replace diagonal and adjacent bands with 0 (Theiler window of 1)

#> 10 x 10 Matrix of class "dgeMatrix"
#>               [,1]         [,2]       [,3]       [,4]       [,5]         [,6]
#>  [1,]  0.000000000  0.000000000 -1.4000435 -1.4000435 -1.4000435 -1.400043517
#>  [2,]  0.000000000  0.000000000  0.0000000  0.2553171  0.2553171  0.255317055
#>  [3,] -2.437263611  0.000000000  0.0000000  0.0000000 -2.4372636 -2.437263611
#>  [4,] -0.005571287 -0.005571287  0.0000000  0.0000000  0.0000000 -0.005571287
#>  [5,]  0.621552721  0.621552721  0.6215527  0.0000000  0.0000000  0.000000000
#>  [6,]  1.148411606  1.148411606  1.1484116  1.1484116  0.0000000  0.000000000
#>  [7,] -1.821817661 -1.821817661 -1.8218177 -1.8218177 -1.8218177  0.000000000
#>  [8,] -0.247325302 -0.247325302 -0.2473253 -0.2473253 -0.2473253 -0.247325302
#>  [9,] -0.244199607 -0.244199607 -0.2441996 -0.2441996 -0.2441996 -0.244199607
#> [10,] -0.282705449 -0.282705449 -0.2827054 -0.2827054 -0.2827054 -0.282705449
#>               [,7]         [,8]         [,9]        [,10]
#>  [1,] -1.400043517 -1.400043517 -1.400043517 -1.400043517
#>  [2,]  0.255317055  0.255317055  0.255317055  0.255317055
#>  [3,] -2.437263611 -2.437263611 -2.437263611 -2.437263611
#>  [4,] -0.005571287 -0.005571287 -0.005571287 -0.005571287
#>  [5,]  0.621552721  0.621552721  0.621552721  0.621552721
#>  [6,]  0.000000000  1.148411606  1.148411606  1.148411606
#>  [7,]  0.000000000  0.000000000 -1.821817661 -1.821817661
#>  [8,]  0.000000000  0.000000000  0.000000000 -0.247325302
#>  [9,] -0.244199607  0.000000000  0.000000000  0.000000000
#> [10,] -0.282705449 -0.282705449  0.000000000  0.000000000