Deprecated functions in package casnet.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("-deprecated").

dyn_comp(df, col_first, col_last, win = NROW(df), fromScaleMin = NULL,
  fromScaleMax = NULL, toScaleMin = NULL, toScaleMax = NULL,
  rescale = TRUE, doPlot = FALSE)

DC_F(df, win, xmin, xmax, col_first, col_last)

DC_D(df, win, xmin, xmax, col_first, col_last)

crit_in(df, win)

dyn_comp

For dyn_comp, use dc_win.

DC_F

For DC_F, use dc_f.

DC_D

For DC_D, use dc_d.

crit_in

For crit_in, use dc_ccp.