Find optimal (C)RQA parameters

A wrapper for various algorithms used to find optimal value pair for the embedding delay and the number of embedding dimensions

crqa_parameters(y, emLag = NULL, maxLag = floor(length(y)/(maxDim +
  1)), lagMethods = c("first.minimum", "global.minimum", "max.lag"),
  estimateDimensions = "preferSmallestInLargestHood", maxDim = 10,
  nnSizes = c(2, 5, 10, 15), nnRadius = 5, nnThres = 10,
  theiler = 0, diagPlot = TRUE, silent = TRUE, ...)

Arguments

y
A numeric vector or time series

emLag
Optimal embedding lag (delay), e.g., provided by an optimising algorithm. If NULL the lags based on the mutual information in lagMethods will be reported. If a numeric value representing a valid lag is passed, this value will be used to estimate the number of dimensions (default = NULL)

maxLag
Maximum embedding lag to consider. Default value is: floor(length(y)/(maxDim+1))

lagMethods
A character vector with one or more of the following strings: "first.minimum","global.minimum","max.lag". If emLag represents a valid lag this value will be reported as "user.lag" (default = c("first.minimum","global.minimum","max.lag"))

estimateDimensions
Decide on an optimal embedding dimension relative to the values in maxDim and lagMethods, according to a number of preferences passed as a character vector. The order in which the preferences appear in the vector affects the selection procedure, with index 1 being most important preference. The following options are available:

preferNone - No optimal number will be picked all other preferences will be ignored

preferSmallestDim - Pick smallest number of dimensions associated with a percentage NN below nnThres

preferSmallestNN - Pick the number of dimensions that is associated with the smallest percentage NN below nnThres

preferSmallestLag - If the value of nnThres does not lead to a unique preference for a pair of dimension and lag values, use the pair with the smallest lag

preferSmallestInLargestHood - The default option: If no unique pair can be found, prefer pairs with smallest values for lag, dimensions, percentage NN for the largest NN size

maxDim
Maximum number of embedding dimensions to consider (default = 10)

nnSizes
Points whose distance is nnSize times further apart than the estimated size of the attractor will be declared false neighbours. See the argument atol in FNN (default = c(2,5,10,15))

nnRadius
If the ratio of the distance between two points in successive dimensions is larger than nnRadius, the points are declared false neighbours. See the argument rtol in FNN (default = 5)

nnThres
Threshold value representing the percentage of Nearest Neighbours that would be acceptable when using N surrogate dimensions. The smallest number of surrogate dimensions that yield a value below the threshold will be considered optimal (default = 10)

theiler
Theiler window on distance matrix (default = 0)

diagPlot
Produce a diagnostic plot the results (default = TRUE)

silent
Silent-ish mode

...
Other parameters passed to timeLag

Value

A list object containing the optimal values (as indicated by the user) and iteration history.

Details

A number of functions are called to determie optimal parameters for delay embedding a time series:

Embedding lag (\(\tau\), emLag): The default is to call est_emLag, which is a wrapper around timeLag with technique="ami" to get lags based on the mutual information function.

Embedding dimension (m, emDim): The default is to call est_emDim, which is a wrapper around FNN

See also