Distance to binary matrix

Distance matrix to binary matrix based on threshold value

Usage

mat_di2bi(distmat, emRad = NA, convMat = FALSE)

Arguments

distmat

Distance matrix

emRad

The radius or threshold value

convMat

Should the matrix be converted from a distmat object of class Matrix::Matrix() to base::matrix() (or vice versa)

Value

A (sparse) matrix with only 0s and 1s

See also