Distance 2 weighted matrix

Distance matrix to weighted matrix based on threshold value

mat_di2we(distmat, emRad, theiler = 0, convMat = FALSE)

Arguments

distmat
Distance matrix

emRad
The radius or threshold value

theiler
Use a theiler window around the line of identity / synchronisation to remove high auto-correlation at short time-lags (default = 0)

convMat
convMat Should the matrix be converted from a distmat object of class Matrix::Matrix() to base::matrix() (or vice versa)

Value

A matrix with 0s and values < threshold distance value

See also