Plot windowed Multiplex Recurrence Network measures

Plot windowed Multiplex Recurrence Network measures

plotMRN_win(
  df_mrn,
  measures = "mi",
  mrnWeights = "mi",
  showSeries = FALSE,
  doPlot = TRUE
)

Arguments

df_mrn

Output from function mrn() with arguments set for a windowed analysis.

measures

Character vector indicating which measures should be plotted. Valid elements in the vector are: "mi" for inter-layer mutual information,"eo" for edge overlap, or any function name from package igraph that can be applied to extract vertex-based measures from a weighted network. This is the multiplex recurrence network with a number of nodes equal to the number of layers and edge weights set by argument mrnWeights. (default = "mi")

mrnWeights

Which measure should be used for the MRN edge-weights? Valid options are "mi" for inter-layer mutual information,"eo" for edge overlap. (default = "mi")

showSeries

Should the time series that constitute the layers be plotted below the windowed MRN measures? This will only work if the mrn() was called using graphs that were generated by function rn() with returnGraph = TRUE, in which case they will have a property V(g)$y1 (and if applicable V(g)$y2), which will be plotted. If more than 1 measure is requested in measures and showSeries = TRUE a seperate plot for each measure will be produced. (default = FALSE)

doPlot

Plot the igraph object.

Value

a ggplot object

See also