Profile Plot

Plot a profile (values of the dimensions) for each phase.

Usage

plotRN_phaseProfiles(
  phaseOutput,
  plotCentroid = NA,
  showEpochLegend = TRUE,
  epochColours = NULL,
  epochLabel = "Phase",
  excludeTransients = FALSE,
  excludePhaseNeighbours = FALSE,
  excludeNonrecurring = TRUE,
  excludeSingularities = TRUE,
  excludeVars = "",
  excludePhases = "",
  showPhaseSize = TRUE,
  returnGraph = FALSE,
  colOrder = NA
)

Arguments

phaseOutput

Output from function rn_phaseInfo

excludeTransients

Should the category "Transient" be excluded from plots? (default = FALSE)

excludeNonrecurring

Should the category "Nonrecurring" be excluded from plots? (default = TRUE)

excludeSingularities

Should the category "Singularity" be excluded from plots? (default = TRUE)

excludeVars

Exclude specific dimension variables by name. Leave empty to include all variables (default = "")

excludePhases

Exclude Phases by their name (variable phase_name). Leave empty to include all Phases (after the other exclusion arguments) (default = "")

showPhaseSize

Show the number of states in each phase in the labels (default = TRUE)

Value

plot

Examples

RN <- rn(y1 = data.frame(x = rnorm(100), y= rnorm(100)), weighted = TRUE)
#> Set `weightedBy` to 'si' due to the value of `weighted`
phase_out <- rn_phaseInfo(RN, returnCentroid = "mean.sd")
#> 
#> ~~~o~~o~~casnet~~o~~o~~~
#> 
#> Recurring states with high similarity will be considered a phase:
#> 
#> selectionMethod = degree
#> maxPhases = 100
#> minStatesinPhase = 2
#> maxStatesinPhase = 100
#> inverseWeight = TRUE
#> 
#> Recurring states will be grouped as a 'Phase ##', as transients (unstable phases) 'Transient ##', phase neighbours (state connects to more than 1 identified phase) 'PhaseNH ##', singularities (state recurs less often than minStatesinPhase) 'Singularity', or as not recurring 'Nonrecurring'.
#> 
#> ~~~o~~o~~casnet~~o~~o~~~
#> Error in dplyr::filter(tmp_stray, !is.na(phase_num)): ℹ In argument: `!is.na(phase_num)`.
#> Caused by error:
#> ! object 'phase_num' not found
plotRN_phaseProfiles(phaseOutput = phase_out)
#> Error: object 'phase_out' not found