Mutliplex Recurrence Network (DEPRECATED)

This function will create a Multiplex Recurrence Network from a list of igraph objects that can be considered the layers of a network. The layers must have the same number of nodes. There are two modes of operation: Layer similarity (MRNweightedBy is set to "InterLayerMI" or "EdgeOvelap") and Layer importance (MRNweightedBy is "RankorderDC"). The former generates weighted MRN based on Interlayer Mutual Information or Edge Overlap, the latter examines the relative importance of each layer by assigning a rank to each vertex (time point), based on a vertex measure passed in argument MRNrankedBy.

rn_multiplex(
  layers,
  MRNweightedBy = c("InterlayerMI", "Edgeoverlap", "RankDC")[1],
  MRNrankedBy = "degree",
  win = NA,
  step = NA,
  overlap = NA,
  alignment = "l",
  doPlot = FALSE,
  doSave = FALSE,
  coord = NA,
  RNnodes = FALSE,
  scaleVertexSize = c(0.01, 5),
  vertexColour = getColours(length(layers)),
  vertexBorderColour = "black",
  showVertexLegend = TRUE,
  showSizeLegend = FALSE,
  alphaV = 0.7,
  scaleEdgeSize = 1/5,
  alphaE = 0.5,
  showEdgeColourLegend = FALSE,
  curvature = -0.6,
  createAnimation = FALSE,
  useImageMagick = FALSE,
  loopAnimation = TRUE,
  transitionLength = 3,
  stateLength = 1,
  gifWidth = 600,
  gifRes = 150,
  noParts = TRUE,
  imageDir = NA,
  silent = TRUE
)

Arguments

layers
A list of igraph objects representing the layers of the multiplex network. The layer networks must all have the same number of vertices.

MRNweightedBy
The measure to be used to evaluate the average structural similarities between the layers of the network. Valid options are: "InterLayerMI" (Mutual information based on similarity of the vertex degree across layers), "EdgeOverlap" (proportion of vertices sharing the same edges across layers), "RankorderDC" (Dynamic Complexity of the inter layer vertex rank order based on the vertex property/measure in MRNrankedBy). Choosing "InterLayerMI" or "EdgeOverlap" will decide which measure is displayed in the plot of the Multiplex RN, both measures will always be returned in the numerical output.

MRNrankedBy
If MRNweightedBy = "RankorderDC", then MRNrankedBy must be a valid igraph command that returns vertex properties, for example: "degree", "strength","hub_score","centr_degree", "transitivity", "betweenness". The appropriate measure type, e.g. for "directed", or "weighted" graphs, will be inferred from the graph properties of the 1st graph object in the layers list. For best results with weighted measures, assign a value to E(g)$weight for each g in layers. (default = "degree")

win
The window size passed to ts_window() in which to evaluate "InterLayerMI" or "EdgeOvelap". If MRNweightedBy = "RankorderDC", it will be the size of the right aligned window in which Dynamic Complexity will be computed using dc_win() (default = NA)

step
The stepsize for the sliding window (default = NA)

overlap
The window overlap passed to ts_window() if MRNweightedBy is "InterLayerMI" or "EdgeOvelap". The value of step will be ignored if overlap is not NA. (default = NA).

alignment
Alignment of the window "l","c","r".

doPlot
Plot the multiplex recurrence network (default = TRUE).

RNnodes
Should the vertices represent the RN of the layers? This is recommended only for a small numbers of vertices. (default = `FALSE``)

scaleVertexSize
Scale the size of the vertices by setting a range for ggplot2::scale_size(). This will not affect the numbers on the size legend (default = c(1,6))

vertexColour
A vector of colours for the vertices. If this is a named list, names will be displayed in the legend.

vertexBorderColour
Draw a border around the vertices. Pass NULL to use the same colour as the fill colour (default = "black")

showVertexLegend
Show the vertex colour legend?

showSizeLegend
Should a legend be shown for the size of the nodes? (default = FALSE)

scaleEdgeSize
Scale the size of the edges by a constant: E(g)$width * scaleEdgeSize (default = 1/5)

showEdgeColourLegend
Should a legend be shown for the colour of the edges? (default = FALSE)

curvature
The curvature parameter for edges see geom_curve() (default = -0.7)

createAnimation
If createAnimation = TRUE and doPlot = TRUE and a windowed analysis is conducted, an animation will be produced using either package gganimate (if useImageMagick = FALSE) or animation (if useImageMagick = FALSE). The main difference is that gganimate has nice animation transition features, but plots the MRN using ggplot2, which does not have great options for displaying the nodes as images. With package animation a sequence of igraph plots will be converted to an animation. If doSave = TRUE the animation will be saved in imageDir as an animated gif by calling either gganimate::anim_save(), or animation::saveGIF() (default = FALSE)

useImageMagick
Should ImageMagick be used to create the animation. NOTE: ImageMagick has to be installed on your system, see animation::saveGIF() (default = FALSE)

loopAnimation
Should the animation loop? (default = TRUE)`

transitionLength
Length of each transition in the animation, ignored if useImageMagick = TRUE (default = 3)

stateLength
Value of state_length if gganimate is used, or the interval in seconds for animation::ani.pause() (default = 1)

gifWidth
Width of the animated gif in pixels. The default width will be 600/150 = 4 in or 10.16 cm (default = 600)

gifRes
Resolution of the animated gif in ppi (default =150)

noParts
Do not plot the individual graphs that make up the animation to the current dev (default = TRUE)

imageDir
Directory to save the layer images and windowed MRN plots. If NA, the value returned by getwd() will be used, if NULL no windowed images will be saved (default = NA)

silent
Silent-ish mode

windowedWeights
If a windowed analysis is conducted and the edges of the graphs in layers have a weight property, there are a number of different ways to handle the weights depending on the value of windowedWeights: "none", "local", and "cumulative". Value "none" will ignore the weights, "local" will limit the range of edges to those edges connecting the vertices contained within the window, "cumulative" will consider all edges connecting to vertices in the current window, including edges from vertices connecting from previous windows. (default = none)

coords
A data frame with layout coordinastes generated by calling any of the igraph layout functions. If NA a circle layout will; be generated (default = NA)

Value

A matrix with edge weights between layers that represent the measure MRNweightedBy.

Examples