Extract Phases from weighted RN

This function will extract phases (regions of attraction in phase space) based on a weighted RN object created with function rn. The assumption is that coordinates in state space that are either close in terms of distance, or are re-visited with short recurrence times, or with high-frequency, are regions of attraction for the system.

rn_phases(
  RN,
  maxPhases = 5,
  minStatesinPhase = 1,
  maxStatesinPhase = NROW(RN),
  cleanUp = TRUE,
  removeSingularities = FALSE,
  standardise = c("none", "mean.sd", "median.mad", "unit")[4],
  returnGraph = FALSE,
  doPhasePlot = FALSE,
  plotCentroid = c("no", "mean.sd", "median.mad", "centroid")[1],
  colOrder = FALSE,
  phaseColours = NULL,
  doSpiralPlot = FALSE,
  doPhaseSeriesPlot = FALSE,
  showEpochLegend = TRUE,
  epochColours = NULL,
  epochLabel = "Phase",
  excludeOther = FALSE,
  excludeNorec = TRUE
)

Arguments

RN
A matrix produced by the function rn

maxPhases
The maximum number of phases to extract. If NA, the value will be set to NROW(RN) (default = 5)

minStatesinPhase
A parameter applied after the extraction of phases (limited by maxPhases). If any extracted phases do not have a minimum number of minStatesinPhase + 1 (= the state that was selected based on node strength), the phase will be removed from the result (default = 1)

maxStatesinPhase
A parameter applied after the extraction of phases (limited by maxPhases). If any extracted phases exceeds a maximum number of maxStatesinPhase + 1 (= the state that was selected based on node strength), the phase will be removed from the result (default = NROW(RN))

cleanUp
Try to assign states to phases that were not assigned by the algorithm. If FALSE, these states will be added to category "Other" (default = TRUE)

removeSingularities
Will remove states that recur only once (nodes with degree(g) == 1) (default = FALSE)

standardise
Standardise by the series using ts_standardise() with adjustN = FALSE (default = "mean.sd")

returnGraph
Return the graph object objects of the plots that have been produced (default = FALSE)

doPhasePlot
Produce a plot with the extracted phases (default = TRUE)

plotCentroid
Currently plots the Mean and SD of the dimensions of the states in each phase over the phase plot (default = FALSE)

colOrder
Should the order of the dimensions reflect the order of the columns in the dataset? If FALSE the order will be based on the values of the dimensions observed in the first extracted phase (default = FALSE)

phaseColours
Colours for the different phases in the phase plot. If epochColours also has a value, phaseColours will be used instead (default = NULL)

doSpiralPlot
Produce a plot of the recurrence network with the nodes coloured by phases (default = FALSE)

doPhaseSeriesPlot
Produce a time series of the phases as they occur with a marginal histogram of their frequency (default = FALSE)

showEpochLegend
Should a legend be shown for the epoch colours? (default = TRUE)

epochColours
A vector of length vcount(g) with colour codes (default = NULL)

epochLabel
A title for the epoch legend (default = "Epoch")

excludeOther
Should the phase "Other" be excluded from plots? (default = FALSE)

excludeNorec
Should the category "No recurrence" be excluded from plots? (default = TRUE)

Value

A data frame with information about the phases.

Details

The method used for the identification of phases is on the properties of the RN object:

If weighted by distance "si", the inverse distance will be used, which means higher weights correspond to closer states.

If weighted by recurrence time "rt", the inverse time will be used, which means higher weights correspond to faster recurrence times.

The procedure is as follows:

Identify the node with the highest strength

Identify the nodes that connect to this node

Identify the node with highest strength that does not connect to the node identified in step 1.

Repeat until criteria set in maxPhases, minStatesinPhase and maxStatesinPhase are triggered.

Clean up. If cleanUp = TRUE the remaining states that connect to nodes of one of the identified phases, but not to the node with highest strength identified in step 1 will be added to that phase. Otherwise these nodes will end up in category "Other".

See also

Examples

# Use the ManyAnalysts dataset to create a phase plot with default settings
data("manyAnalystsESM")
df <- manyAnalystsESM[4:10]
RN <- rn(y1 = df, doEmbed = FALSE, weighted = TRUE, weightedBy = "si", emRad = NA)

#> <sparse>[ <logic> ] : .M.sub.i.logical() maybe inefficient

#> 
#> Searching for a radius that will yield 0.05 for RR 
#> Iteration 1 
#> Iteration 2 
#> Iteration 3 
#> Iteration 4 
#> Iteration 5 
#> Iteration 6 
#> Iteration 7 
#> Iteration 8 
#> Iteration 9 

#> 
#> Converged! Found an appropriate radius...

# This returns 6 phases which have minimally 1 state
rn_phases(RN, maxPhases = 10, doPhasePlot = TRUE)

#> Error: Problem with `mutate()` column `phase_size`.
#> ℹ `phase_size = n()`.
#> x could not find function "n"
#> ℹ The error occurred in group 1: phase_name = "No recurrence".

# Use min. number of states as the extraction criterion
rn_phases(RN, maxPhases = NA, minStatesinPhase = 7, doPhasePlot = TRUE)

#> Error: Problem with `mutate()` column `phase_size`.
#> ℹ `phase_size = n()`.
#> x could not find function "n"
#> ℹ The error occurred in group 1: phase_name = "No recurrence".