Recurrence Time Scaleogram

Display a recurrence network in a space representing Time (x-axis) x Scale (y-axis). The scale axis will be determined by the latency between the occurence of a value in the (embedded) time series vector and its recurrences in the future (i.e. only the upper triangle of the recurrence matrix will be displayed, excluding the diagonal).

rn_scaleoGram(RN, returnOnlyObject = FALSE)

Arguments

RN
A thresholded recurrence matrix generated by function rn()

returnOnlyObject
Return the ggplot / ggraph object only, do not draw the plot (default = FALSE)

Value

A ggraph graph object

See also