rp_nzdiags

Get all nonzero diagonals of a binary matrix, or, diagonals specified as a vector by argument d.

rp_nzdiags(
  RM = NULL,
  d = NULL,
  returnVectorList = TRUE,
  returnNZtriplets = FALSE,
  removeNZ = TRUE,
  silent = TRUE
)

Arguments

RM

A binary (0,1) matrix.

d

An optional vector of diagonals to extract.

returnVectorList

Return list

returnNZtriplets

Return a dataframe with coordinates of only nonzero elements in diagonals (default = FALSE)

removeNZ

Remove nonzero diagonals if TRUE. If FALSE returns the full diagonals matrix. Use e.g. to plot diagonal recurrence profiles (default = TRUE)

silent

Silent-ish mode

Value

A matrix object with nonzero diagonals as columns and/or a dataframe with coordinates of nonzero diagonal elements

See also

Author

Fred Hasselman