Tidy eval helpers

rlang::sym() creates a symbol from a string and rlang::syms() creates a list of symbols from a character vector.

rlang::expr() and rlang::quo() quote one expression. quo() wraps the quoted expression in a quosure.

The plural variants rlang::exprs() and rlang::quos() return a list of quoted expressions or quosures.

rlang::enexpr() and rlang::enquo() capture the expression supplied as argument by the user of the current function (enquo() wraps this expression in a quosure).

rlang::enexprs() and rlang::enquos() capture multiple expressions supplied as arguments, including ....

exprs() is not exported to avoid conflicts with Biobase::exprs(), therefore one should always use rlang::exprs().

To learn more about tidy eval and how to use these tools, visit http://rlang.r-lib.org and the Metaprogramming section of Advanced R.