Find change indices

Find change indices

ts_changeindex(
  y,
  returnRectdata = TRUE,
  groupVar = NULL,
  labelVar = NULL,
  discretize = FALSE,
  nbins = 5
)

Arguments

y
An indicator variable representing different levels of a variable or factor

returnRectdata
Return a dataframe suitable for shading a ggplot2 graph with ggplot2::geom_rect()

groupVar
Pass a value (length 1) or variable (length of y) that can be used as a variable to join the indices by if returnRectdata = TRUE

labelVar
If y is not a character vector, provide a vector of labels equal to length(y)

discretize
If y is a continuous variable, setting discretize = TRUE will partition the values of y into nbins number of bins, each value of y will be replaced by its bin number.

nbins
Number of bins to use to change a continuous y (if discretize = TRUE) into a variable with nbins levels

Value

Either a vector with the indices of change in y, or, a data frame with variables xmin,xmax,ymin,ymax,label

See also