Adjust time series by summation order

Many fluctuation analyses assume a time series' Hurst exponent is within the range of 0.2 - 1.2. If this is not the case it is sensible to make adjustments to the time series, as well as the resutling Hurst exponent.

ts_sumorder(y, scaleS = NULL, polyOrder = 1, dataMin = 4)

Arguments

y

A time series of numeric vector

scaleS

The scales to consider for DFA1

polyOrder

Order of polynomial for detrending in DFA (default = 1)

dataMin

Minimum number of data points in a bin needed to calculate detrended fluctuation

Value

The input vector, possibly adjusted based on H with an attribute "Hadj" containing an integer by which a Hurst exponent calculated from the series should be adjusted.

Details

Following recommendations by https://www.frontiersin.org/files/Articles/23948/fphys-03-00141-r2/image_m/fphys-03-00141-t001.jpgIhlen (2012), a global Hurst exponent is estimated using DFA and y is adjusted accordingly:

1.2 < H < 1.8 first derivative of y, atribute Hadj = 1

H > 1.8 second derivative of y, atribute Hadj = 2

H < 0.2 y is centered and integrated, atribute Hadj = -1

0.2 <= H <= 1.2 y is unaltered, atribute Hadj = 0

References

Ihlen, E. A. F. E. (2012). Introduction to multifractal detrended fluctuation analysis in Matlab. Frontiers in physiology, 3, 141.

See also