Replace matrix diagonals

Sets a band of matrix diagonals to any given value

bandReplace(mat, lower, upper, value = NA, silent = TRUE)

Arguments

mat

A Matrix

lower

Lower diagonal to be included in the band (should be \(\le 0\))

upper

Upper diagonal to be included in the band (should be \(\ge 0\))

value

A single value to replace all values in the selected band (default = NA)

silent

Operate in silence, only (some) warnings will be shown (default = TRUE)

Value

A matrix in which the values in the selected diagonals have been replaced

See also

Author

Fred Hasselman

Examples

# Create a 10 by 10 matrix
library(Matrix)
m <- Matrix(rnorm(10),10,10)

bandReplace(m,-1,1,0)   # Replace diagonal and adjacent bands with 0 (Theiler window of 1)
#> 10 x 10 sparse Matrix of class "dgCMatrix"
#>                                                                        
#>  [1,]  .          .          0.3009553  0.3009553  0.3009553  0.3009553
#>  [2,]  .          .          .          0.0525553  0.0525553  0.0525553
#>  [3,]  0.2504568  .          .          .          0.2504568  0.2504568
#>  [4,] -0.2834756 -0.2834756  .          .          .         -0.2834756
#>  [5,] -0.1210398 -0.1210398 -0.1210398  .          .          .        
#>  [6,]  1.2333389  1.2333389  1.2333389  1.2333389  .          .        
#>  [7,] -0.3617100 -0.3617100 -0.3617100 -0.3617100 -0.3617100  .        
#>  [8,] -1.2115548 -1.2115548 -1.2115548 -1.2115548 -1.2115548 -1.2115548
#>  [9,] -0.5660410 -0.5660410 -0.5660410 -0.5660410 -0.5660410 -0.5660410
#> [10,]  0.4767876  0.4767876  0.4767876  0.4767876  0.4767876  0.4767876
#>                                                  
#>  [1,]  0.3009553  0.3009553  0.3009553  0.3009553
#>  [2,]  0.0525553  0.0525553  0.0525553  0.0525553
#>  [3,]  0.2504568  0.2504568  0.2504568  0.2504568
#>  [4,] -0.2834756 -0.2834756 -0.2834756 -0.2834756
#>  [5,] -0.1210398 -0.1210398 -0.1210398 -0.1210398
#>  [6,]  .          1.2333389  1.2333389  1.2333389
#>  [7,]  .          .         -0.3617100 -0.3617100
#>  [8,]  .          .          .         -1.2115548
#>  [9,] -0.5660410  .          .          .        
#> [10,]  0.4767876  0.4767876  .          .