Elastic Scaler - A Flexible Rescale Function

The 'elastic scaler'will rescale numeric vectors (1D, or columns in a matrix or data.frame) to a user defined minimum and maximum, either based on the extrema in the data, or, a minimum and maximum defined by the user.

elascer(x, mn = NA, mx = NA, lo = 0, hi = 1, groupwise = FALSE)

Arguments

x
Input vector or data frame.

mn
Minimum value of original, defaults to min(x, na.rm = TRUE) if set to NA.

mx
Maximum value of original, defaults to max(x, na.rm = TRUE) if set to NA.

lo
Maximum value to rescale to, defaults to 1.

hi
Minimum value to rescale to, defaults to 0.

groupwise
If x is a data frame with 2+ columns, mn = NA and/or mx = NA and groupwise = TRUE, scaling will occur

Value

scaled inout

Details

Three uses:

elascer(x) - Scale x to data range: min(x.out)==0; max(x.out)==1

elascer(x,mn,mx) - Scale x to arg. range: min(x.out)==mn==0; max(x.out)==mx==1

elascer(x,mn,mx,lo,hi) - Scale x to arg. range: min(x.out)==mn==lo; max(x.out)==mx==hi

Examples

# Works on numeric objects
somenumbers <- cbind(c(-5,100,sqrt(2)),c(exp(1),0,-pi))

elascer(somenumbers)

#>           V1         V2
#> 1 0.00000000 0.07350745
#> 2 1.00000000 0.04761905
#> 3 0.06108775 0.01769912

elascer(somenumbers,mn=-100)

#>          V1        V2
#> 1 0.4750000 0.5135914
#> 2 1.0000000 0.5000000
#> 3 0.5070711 0.4842920

# Values < mn will return < lo (default=0)
# Values > mx will return > hi (default=1)
elascer(somenumbers,mn=-1,mx=99)

#>            V1          V2
#> 1 -0.04000000  0.03718282
#> 2  1.01000000  0.01000000
#> 3  0.02414214 -0.02141593

elascer(somenumbers,lo=-1,hi=1)

#>           V1         V2
#> 1 -1.0000000 -0.8529851
#> 2  1.0000000 -0.9047619
#> 3 -0.8778245 -0.9646018

elascer(somenumbers,lo=-1,hi=1, groupwise = TRUE)

#>           V1          V2
#> 1 -1.0000000  1.00000000
#> 2  1.0000000  0.07223889
#> 3 -0.8778245 -1.00000000

elascer(somenumbers,mn=-10,mx=100,lo=-1,hi=4)

#>           V1         V2
#> 1 -0.7727273 -0.4218963
#> 2  4.0000000 -0.5454545
#> 3 -0.4811721 -0.6882542

elascer(somenumbers,mn= NA,mx=100,lo=-1,hi=4, groupwise = TRUE)

#>           V1         V2
#> 1 -1.0000000 -0.7159306
#> 2  4.0000000 -0.8477049
#> 3 -0.6945613 -1.0000000