Standardised Dispersion Analysis (SDA).

fd_sda

fd_sda(
  y,
  fs = NULL,
  standardise = c("mean.sd", "median.mad")[1],
  detrend = FALSE,
  polyOrder = 1,
  adjustSumOrder = FALSE,
  scaleMin = 2,
  scaleMax = floor(log2(NROW(y)/2)),
  scaleResolution = 30,
  scaleS = NA,
  overlap = 0,
  minData = 4,
  doPlot = FALSE,
  returnPlot = FALSE,
  returnPLAW = FALSE,
  returnInfo = FALSE,
  silent = FALSE,
  noTitle = FALSE,
  tsName = "y"
)

Arguments

y
A numeric vector or time series object.

fs
Sample rate (default = NULL)

standardise
standardise the series (default = "mean.sd")

detrend
Subtract linear trend from the series (default = FALSE)

polyOrder
Order of detrending polynomial

adjustSumOrder
Adjust the time series (summation or differencing), based on the global scaling exponent, see e.g. https://www.frontiersin.org/files/Articles/23948/fphys-03-00141-r2/image_m/fphys-03-00141-t001.jpgIhlen (2012) (default = FALSE)

scaleMin
Minimium scale to use

scaleMax
Maximum scale to use

scaleResolution
The scales at which the standardised fluctuations are calculated as: (scaleMax-scaleMin)/scaleResolution

scaleS
If not NA, it should be a numeric vector listing the scales on which to evaluate the fluctuations. Arguments scaleMax, scaleMin, scaleResolution will be ignored.

overlap
Turn SDA into a sliding window analysis. A number in [0 ... 1] representing the amount of 'bin overlap'. If length(y) = 1024 and overlap is .5, a scale of 4 will be considered a sliding window of size 4 with stepsize floor(.5 * 4) = 2 (default = 0)

minData
Minimum number of data points in a bin needed to calculate standardised dispersion

doPlot
Output the log-log scale versus fluctuation plot with linear fit by calling function plotFD_loglog() (default = TRUE)

returnPlot
Return ggplot2 object (default = FALSE)

returnPLAW
Return the power law data (default = FALSE)

returnInfo
Return all the data used in SDA (default = FALSE)

silent
Silent-ish mode

noTitle
Do not generate a title (only the subtitle)

tsName
Name of y added as a subtitle to the plot

Value

A list object containing:

A data matrix PLAW with columns freq.norm, size and bulk.

Estimate of scaling exponent sap based on a fit over the standard range (fullRange), or on a user defined range fitRange.

Estimate of the the Fractal Dimension (FD) using conversion formula's reported in Hasselman(2013).

Information output by various functions.

References

Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. http://doi.org/10.3389/fphys.2013.00075

See also