Mutual Information Function

Calculate the lagged mutual information fucntion within (auto-mif) or between (cross-mif) time series, or, conditional on another time series (conditional-cross-mif). Alternatively, calculate the total information of a multivariate dataset for different lags.

mif(y, lags = -10:10, nbins = ceiling(2 * NROW(y)^(1/3)),
  doPlot = FALSE, surTest = FALSE)

Arguments

y
A Nx1 matrix for auto-mif, a Nx2 matrix or data frame for cross-mif, a Nx3 matrix or data frame for mif between col 1 and 2 conditional on col 3; or a NxM matrix or data frame for the multi-information function. Mutual information for each lag will be calculated using functions in package infotheo::infotheo() for lags lagged versions of the time series.

lags
The lags to evaluate mutual information.

nbins
The number of bins passed to infotheo::discretize() if y is a matrix or casnet::ts_discrete()

doPlot
Produce a plot of the symbolic time series (default = FALSE)

surTest
Either FALSE or an alpha level for conducting a test of significance using simple surrogates, e.g. surTes = .05. The surrogates will be created from the transition probabilities of the discretised time series, i.e. the probability of observing bin j when the current value is in bin j .

Value

The auto- or cross-mi function

See also

Examples

# Lags to evaluate mututal information
lags <- -10:30

# Auto-mutual information
y1 <- sin(seq(0, 100, by = 1/8)*pi)

(mif(data.frame(y1),lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.653815 1.927735 1.646619 1.573201 1.567816 1.678095 1.455365 1.497499 
#>       -2       -1        0        1        2        3        4        5 
#> 1.476553 2.868931 2.868931 2.868931 1.476553 1.497499 1.455365 1.678095 
#>        6        7        8        9       10       11       12       13 
#> 1.567816 1.573201 1.646619 1.927735 1.653815 1.550633 1.923477 1.562827 
#>       14       15       16       17       18       19       20       21 
#> 1.607086 1.621646 1.558334 1.859470 1.473869 1.623019 1.476240 1.515003 
#>       22       23       24       25       26       27       28       29 
#> 1.453358 1.751048 1.566279 1.964346 1.604698 1.668617 1.568317 1.564823 
#>       30 
#> 1.734763 
#> attr(,"miType")
#> [1] "I(X;X)"

# Cross-mututal information, y2 is a lagged version y1
y2 <- y1[10:801]

y <- data.frame(ts_trimfill(y1, y2, action = "trim.cut"))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.623019 1.473869 1.859470 1.558334 1.621646 1.607086 1.562827 1.923477 
#>       -2       -1        0        1        2        3        4        5 
#> 1.550633 1.653815 1.653815 1.653815 1.921974 1.687149 1.582642 1.610835 
#>        6        7        8        9       10       11       12       13 
#> 1.694766 1.456779 1.479514 1.468117 2.868171 1.483081 1.499478 1.471427 
#>       14       15       16       17       18       19       20       21 
#> 1.605462 1.578427 1.572016 1.660324 1.909035 1.644244 1.560248 1.923019 
#>       22       23       24       25       26       27       28       29 
#> 1.566571 1.604723 1.613996 1.696857 1.861291 1.474971 1.650919 1.465343 
#>       30 
#> 1.510893 
#> attr(,"miType")
#> [1] "I(X;Y)"

# Conditional mutual information, add some noise to y2 and add it as a 3rd column
y$s <- y2+rnorm(NROW(y2))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.824487 1.744000 1.983888 1.878350 2.000039 1.997520 2.003551 2.057971 
#>       -2       -1        0        1        2        3        4        5 
#> 1.883325 1.902833 1.902833 1.902833 2.040702 1.983402 1.984676 2.005857 
#>        6        7        8        9       10       11       12       13 
#> 2.022351 1.786048 1.787322 1.762830 2.511825 1.857320 1.916479 1.900262 
#>       14       15       16       17       18       19       20       21 
#> 1.934479 1.925594 1.890592 1.891668 2.027751 1.999994 1.979257 2.161102 
#>       22       23       24       25       26       27       28       29 
#> 2.004768 1.966695 1.916750 1.919306 1.983114 1.870310 1.956632 1.897670 
#>       30 
#> 1.925018 
#> attr(,"miType")
#> [1] "I(X;Y|Z)"

# Multi-information, the information of the entire multivariate series at each lag
y$y3 <- cumsum(rnorm(NROW(y)))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 5.011812 4.997671 4.996383 4.996535 5.149310 5.151574 5.143682 5.140343 
#>       -2       -1        0        1        2        3        4        5 
#> 5.137672 5.137097 5.137097 5.137097 5.134198 5.133324 5.138442 5.144145 
#>        6        7        8        9       10       11       12       13 
#> 5.147721 4.987161 4.992326 4.993420 5.006067 5.010936 5.012313 5.018315 
#>       14       15       16       17       18       19       20       21 
#> 5.016186 5.144422 5.140042 5.146757 5.147810 5.152227 5.158326 5.163626 
#>       22       23       24       25       26       27       28       29 
#> 5.160740 5.161417 5.159608 5.164054 5.012035 5.013560 5.025781 5.025573 
#>       30 
#> 5.032717 
#> attr(,"miType")
#> [1] "I(X;Y;Z;...;N)"