Mutual Information Function

Calculate the lagged mutual information fucntion within (auto-mif) or between (cross-mif) time series, or, conditional on another time series (conditional-cross-mif). Alternatively, calculate the total information of a multivariate dataset for different lags.

mif(y, lags = -10:10, nbins = ceiling(2 * NROW(y)^(1/3)),
  doPlot = FALSE, surTest = FALSE)

Arguments

y
A Nx1 matrix for auto-mif, a Nx2 matrix or data frame for cross-mif, a Nx3 matrix or data frame for mif between col 1 and 2 conditional on col 3; or a NxM matrix or data frame for the multi-information function. Mutual information for each lag will be calculated using functions in package infotheo for lags lagged versions of the time series.

lags
The lags to evaluate mutual information.

nbins
The number of bins passed to discretize if y is a matrix or ts_discrete

doPlot
Produce a plot of the symbolic time series (default = FALSE)

surTest
Either FALSE or an alpha level for conducting a test of significance using simple surrogates, e.g. surTes = .05. The surrogates will be created from the transition probabilities of the discretised time series, i.e. the probability of observing bin j when the current value is in bin j .

Value

The auto- or cross-mi function

Examples

# Lags to evaluate mututal information
lags <- -10:30

# Auto-mutual information
y1 <- sin(seq(0, 100, by = 1/8)*pi)

(mif(data.frame(y1),lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.653815 1.927735 1.646619 1.573201 1.567816 1.678095 1.455365 1.497499 
#>       -2       -1        0        1        2        3        4        5 
#> 1.476553 2.868931 2.868931 2.868931 1.476553 1.497499 1.455365 1.678095 
#>        6        7        8        9       10       11       12       13 
#> 1.567816 1.573201 1.646619 1.927735 1.653815 1.550633 1.923477 1.562827 
#>       14       15       16       17       18       19       20       21 
#> 1.607086 1.621646 1.558334 1.859470 1.473869 1.623019 1.476240 1.515003 
#>       22       23       24       25       26       27       28       29 
#> 1.453358 1.751048 1.566279 1.964346 1.604698 1.668617 1.568317 1.564823 
#>       30 
#> 1.734763 
#> attr(,"miType")
#> [1] "I(X;X)"

# Cross-mututal information, y2 is a lagged version y1
y2 <- y1[10:801]

y <- data.frame(ts_trimfill(y1, y2, action = "trim.cut"))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.623019 1.473869 1.859470 1.558334 1.621646 1.607086 1.562827 1.923477 
#>       -2       -1        0        1        2        3        4        5 
#> 1.550633 1.653815 1.653815 1.653815 1.921974 1.687149 1.582642 1.610835 
#>        6        7        8        9       10       11       12       13 
#> 1.694766 1.456779 1.479514 1.468117 2.868171 1.483081 1.499478 1.471427 
#>       14       15       16       17       18       19       20       21 
#> 1.605462 1.578427 1.572016 1.660324 1.909035 1.644244 1.560248 1.923019 
#>       22       23       24       25       26       27       28       29 
#> 1.566571 1.604723 1.613996 1.696857 1.861291 1.474971 1.650919 1.465343 
#>       30 
#> 1.510893 
#> attr(,"miType")
#> [1] "I(X;Y)"

# Conditional mutual information, add some noise to y2 and add it as a 3rd column
y$s <- y2+rnorm(NROW(y2))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.808994 1.715093 1.922049 1.844053 2.001774 2.015249 1.959424 2.050946 
#>       -2       -1        0        1        2        3        4        5 
#> 1.837244 1.863161 1.863161 1.863161 1.981409 1.948618 1.933928 2.028668 
#>        6        7        8        9       10       11       12       13 
#> 2.016915 1.790998 1.752539 1.687093 2.475703 1.796621 1.873929 1.861375 
#>       14       15       16       17       18       19       20       21 
#> 1.903849 1.902230 1.814849 1.826678 1.949864 1.958285 1.926041 2.127642 
#>       22       23       24       25       26       27       28       29 
#> 1.925229 1.909716 1.868772 1.844328 1.916569 1.795793 1.923447 1.861611 
#>       30 
#> 1.871509 
#> attr(,"miType")
#> [1] "I(X;Y|Z)"

# Multi-information, the information of the entire multivariate series at each lag
y$y3 <- cumsum(rnorm(NROW(y)))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 4.995877 4.989931 4.986886 4.988814 5.135218 5.138595 5.137748 5.133324 
#>       -2       -1        0        1        2        3        4        5 
#> 5.130662 5.128346 5.128346 5.128346 5.130693 5.129814 5.130751 5.140627 
#>        6        7        8        9       10       11       12       13 
#> 5.138913 4.985397 4.987028 4.988115 4.994342 4.999196 4.998782 5.002990 
#>       14       15       16       17       18       19       20       21 
#> 5.012191 5.139076 5.141826 5.144970 5.147810 5.148645 5.152946 5.151056 
#>       22       23       24       25       26       27       28       29 
#> 5.157144 5.161417 5.159608 5.162249 5.010228 5.013560 5.020344 5.019444 
#>       30 
#> 5.024764 
#> attr(,"miType")
#> [1] "I(X;Y;Z;...;N)"