Mutual Information Function

Calculate the lagged mutual information fucntion within (auto-mif) or between (cross-mif) time series, or, conditional on another time series (conditional-cross-mif). Alternatively, calculate the total information of a multivariate dataset for different lags.

mif(
  y,
  lags = -10:10,
  nbins = ceiling(2 * NROW(y)^(1/3)),
  doPlot = FALSE,
  surTest = FALSE,
  alpha = 0.05
)

Arguments

y
A Nx1 matrix for auto-mif, a Nx2 matrix or data frame for cross-mif, a Nx3 matrix or data frame for mif between col 1 and 2 conditional on col 3; or a NxM matrix or data frame for the multi-information function. Mutual information for each lag will be calculated using functions in package infotheo::infotheo() for lags lagged versions of the time series.

lags
The lags to evaluate mutual information.

nbins
The number of bins passed to infotheo::discretize() if y is a matrix or ts_discrete()

doPlot
Produce a plot of the symbolic time series by calling plotRED_mif() (default = FALSE)

surTest
If TRUE, a surrogate will be conducted using simple surrogates. The surrogates will be created from the transition probabilities of the discretised time series, i.e. the probability of observing bin j when the current value is in bin j. The number of surrogates needed will be computed based on the value of the alpha parameter, conceived as a one-sided test: mi > 0.

alpha
The alpha level for the surrogate test (default = 0.05)

Value

The auto- or cross-mi function

See also

Examples

# Lags to evaluate mututal information
lags <- -10:30

# Auto-mutual information
y1 <- sin(seq(0, 100, by = 1/8)*pi)

(mif(data.frame(y1),lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.653815 1.927735 1.646619 1.573201 1.567816 1.678095 1.455365 1.497499 
#>       -2       -1        0        1        2        3        4        5 
#> 1.476553 2.868931 2.868931 2.868931 1.476553 1.497499 1.455365 1.678095 
#>        6        7        8        9       10       11       12       13 
#> 1.567816 1.573201 1.646619 1.927735 1.653815 1.550633 1.923477 1.562827 
#>       14       15       16       17       18       19       20       21 
#> 1.607086 1.621646 1.558334 1.859470 1.473869 1.623019 1.476240 1.515003 
#>       22       23       24       25       26       27       28       29 
#> 1.453358 1.751048 1.566279 1.964346 1.604698 1.668617 1.568317 1.564823 
#>       30 
#> 1.734763 
#> attr(,"miType")
#> [1] "I(X;X)"

# Cross-mututal information, y2 is a lagged version y1
y2 <- y1[10:801]

y <- data.frame(ts_trimfill(y1, y2, action = "trim.cut"))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.623019 1.473869 1.859470 1.558334 1.621646 1.607086 1.562827 1.923477 
#>       -2       -1        0        1        2        3        4        5 
#> 1.550633 1.653815 1.653815 1.653815 1.921974 1.687149 1.582642 1.610835 
#>        6        7        8        9       10       11       12       13 
#> 1.694766 1.456779 1.479514 1.468117 2.868171 1.483081 1.499478 1.471427 
#>       14       15       16       17       18       19       20       21 
#> 1.605462 1.578427 1.572016 1.660324 1.909035 1.644244 1.560248 1.923019 
#>       22       23       24       25       26       27       28       29 
#> 1.566571 1.604723 1.613996 1.696857 1.861291 1.474971 1.650919 1.465343 
#>       30 
#> 1.510893 
#> attr(,"miType")
#> [1] "I(X;Y)"

# Conditional mutual information, add some noise to y2 and add it as a 3rd column
y$s <- y2+rnorm(NROW(y2))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.797108 1.713710 1.921446 1.810113 1.970283 1.987413 1.927730 2.069501 
#>       -2       -1        0        1        2        3        4        5 
#> 1.822566 1.878617 1.878617 1.878617 1.994880 1.950518 1.977648 1.978940 
#>        6        7        8        9       10       11       12       13 
#> 1.962697 1.791254 1.740091 1.706107 2.476214 1.794912 1.853147 1.872529 
#>       14       15       16       17       18       19       20       21 
#> 1.915484 1.907470 1.843760 1.831617 1.973862 1.977525 1.955104 2.105902 
#>       22       23       24       25       26       27       28       29 
#> 1.957366 1.962739 1.883271 1.860350 1.925634 1.760896 1.914528 1.852778 
#>       30 
#> 1.857997 
#> attr(,"miType")
#> [1] "I(X;Y|Z)"

# Multi-information, the information of the entire multivariate series at each lag
y$y3 <- cumsum(rnorm(NROW(y)))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 4.986357 4.988830 4.987553 4.991244 5.139406 5.137500 5.138411 5.126967 
#>       -2       -1        0        1        2        3        4        5 
#> 5.127818 5.127256 5.127256 5.127256 5.127849 5.130477 5.134928 5.139531 
#>        6        7        8        9       10       11       12       13 
#> 5.139578 4.987161 4.988794 4.988115 4.995444 4.998527 4.998112 5.004096 
#>       14       15       16       17       18       19       20       21 
#> 5.012627 5.139076 5.140042 5.141397 5.140654 5.139690 5.135012 5.136690 
#>       22       23       24       25       26       27       28       29 
#> 5.140962 5.147014 5.152398 5.155028 5.001190 5.006321 5.013096 5.012186 
#>       30 
#> 5.012045 
#> attr(,"miType")
#> [1] "I(X;Y;Z;...;N)"