Mutual Information Function

Calculate the lagged mutual information fucntion within (auto-mif) or between (cross-mif) time series, or, conditional on another time series (conditional-cross-mif). Alternatively, calculate the total information of a multivariate dataset for different lags.

mif(
  y,
  lags = -10:10,
  nbins = ceiling(2 * NROW(y)^(1/3)),
  doPlot = FALSE,
  surTest = FALSE,
  alpha = 0.05
)

Arguments

y
A Nx1 matrix for auto-mif, a Nx2 matrix or data frame for cross-mif, a Nx3 matrix or data frame for mif between col 1 and 2 conditional on col 3; or a NxM matrix or data frame for the multi-information function. Mutual information for each lag will be calculated using functions in package infotheo::infotheo() for lags lagged versions of the time series.

lags
The lags to evaluate mutual information.

nbins
The number of bins passed to infotheo::discretize() if y is a matrix or ts_discrete()

doPlot
Produce a plot of the symbolic time series by calling plotRED_mif() (default = FALSE)

surTest
If TRUE, a surrogate will be conducted using simple surrogates. The surrogates will be created from the transition probabilities of the discretised time series, i.e. the probability of observing bin j when the current value is in bin j. The number of surrogates needed will be computed based on the value of the alpha parameter, conceived as a one-sided test: mi > 0.

alpha
The alpha level for the surrogate test (default = 0.05)

Value

The auto- or cross-mi function

See also

Examples

# Lags to evaluate mututal information
lags <- -10:30

# Auto-mutual information
y1 <- sin(seq(0, 100, by = 1/8)*pi)

(mif(data.frame(y1),lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.653815 1.927735 1.646619 1.573201 1.567816 1.678095 1.455365 1.497499 
#>       -2       -1        0        1        2        3        4        5 
#> 1.476553 2.868931 2.868931 2.868931 1.476553 1.497499 1.455365 1.678095 
#>        6        7        8        9       10       11       12       13 
#> 1.567816 1.573201 1.646619 1.927735 1.653815 1.550633 1.923477 1.562827 
#>       14       15       16       17       18       19       20       21 
#> 1.607086 1.621646 1.558334 1.859470 1.473869 1.623019 1.476240 1.515003 
#>       22       23       24       25       26       27       28       29 
#> 1.453358 1.751048 1.566279 1.964346 1.604698 1.668617 1.568317 1.564823 
#>       30 
#> 1.734763 
#> attr(,"miType")
#> [1] "I(X;X)"
#> attr(,"lags")
#>  [1] -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1   2   3   4   5   6   7   8
#> [20]   9  10  11  12  13  14  15  16  17  18  19  20  21  22  23  24  25  26  27
#> [39]  28  29  30
#> attr(,"nbins")
#> [1] 19

# Cross-mututal information, y2 is a lagged version y1
y2 <- y1[10:801]

y <- data.frame(ts_trimfill(y1, y2, action = "trim.cut"))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.623019 1.473869 1.859470 1.558334 1.621646 1.607086 1.562827 1.923477 
#>       -2       -1        0        1        2        3        4        5 
#> 1.550633 1.653815 1.653815 1.653815 1.921974 1.687149 1.582642 1.610835 
#>        6        7        8        9       10       11       12       13 
#> 1.694766 1.456779 1.479514 1.468117 2.868171 1.483081 1.499478 1.471427 
#>       14       15       16       17       18       19       20       21 
#> 1.605462 1.578427 1.572016 1.660324 1.909035 1.644244 1.560248 1.923019 
#>       22       23       24       25       26       27       28       29 
#> 1.566571 1.604723 1.613996 1.696857 1.861291 1.474971 1.650919 1.465343 
#>       30 
#> 1.510893 
#> attr(,"miType")
#> [1] "I(X;Y)"
#> attr(,"lags")
#>  [1] -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1   2   3   4   5   6   7   8
#> [20]   9  10  11  12  13  14  15  16  17  18  19  20  21  22  23  24  25  26  27
#> [39]  28  29  30
#> attr(,"nbins")
#> [1] 19

# Conditional mutual information, add some noise to y2 and add it as a 3rd column
y$s <- y2+rnorm(NROW(y2))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.811587 1.741921 1.948223 1.829110 1.957668 1.976480 1.921872 2.029502 
#>       -2       -1        0        1        2        3        4        5 
#> 1.829539 1.841772 1.841772 1.841772 1.981870 1.912914 1.948178 1.984967 
#>        6        7        8        9       10       11       12       13 
#> 1.970052 1.750787 1.753466 1.704676 2.493827 1.824687 1.861592 1.861961 
#>       14       15       16       17       18       19       20       21 
#> 1.886727 1.858509 1.814226 1.845090 1.971008 1.931950 1.905939 2.072356 
#>       22       23       24       25       26       27       28       29 
#> 1.910819 1.917387 1.859806 1.876210 1.981597 1.798201 1.941952 1.860176 
#>       30 
#> 1.875081 
#> attr(,"miType")
#> [1] "I(X;Y|Z)"
#> attr(,"lags")
#>  [1] -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1   2   3   4   5   6   7   8
#> [20]   9  10  11  12  13  14  15  16  17  18  19  20  21  22  23  24  25  26  27
#> [39]  28  29  30
#> attr(,"nbins")
#> [1] 19

# Multi-information, the information of the entire multivariate series at each lag
y$y3 <- cumsum(rnorm(NROW(y)))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 5.006500 4.997671 4.996383 4.993673 5.145355 5.145201 5.142589 5.143853 
#>       -2       -1        0        1        2        3        4        5 
#> 5.136581 5.140598 5.140598 5.140598 5.138365 5.138588 5.135591 5.138436 
#>        6        7        8        9       10       11       12       13 
#> 5.139578 4.993118 4.992993 4.995855 5.001424 5.002742 5.004107 5.002990 
#>       14       15       16       17       18       19       20       21 
#> 5.007960 5.146876 5.144283 5.156363 5.162795 5.167232 5.162590 5.156443 
#>       22       23       24       25       26       27       28       29 
#> 5.155346 5.156016 5.159608 5.165859 5.013842 5.013560 5.023969 5.023073 
#>       30 
#> 5.024764 
#> attr(,"miType")
#> [1] "I(X;Y;Z;...;N)"
#> attr(,"lags")
#>  [1] -10  -9  -8  -7  -6  -5  -4  -3  -2  -1   0   1   2   3   4   5   6   7   8
#> [20]   9  10  11  12  13  14  15  16  17  18  19  20  21  22  23  24  25  26  27
#> [39]  28  29  30
#> attr(,"nbins")
#> [1] 19