Mutual Information Function

Calculate the lagged mutual information fucntion within (auto-mif) or between (cross-mif) time series, or, conditional on another time series (conditional-cross-mif). Alternatively, calculate the total information of a multivariate dataset for different lags.

mif(y, lags = -10:10, nbins = ceiling(2 * NROW(y)^(1/3)),
  doPlot = FALSE, surTest = FALSE, alpha = 0.05)

Arguments

y
A Nx1 matrix for auto-mif, a Nx2 matrix or data frame for cross-mif, a Nx3 matrix or data frame for mif between col 1 and 2 conditional on col 3; or a NxM matrix or data frame for the multi-information function. Mutual information for each lag will be calculated using functions in package infotheo::infotheo() for lags lagged versions of the time series.

lags
The lags to evaluate mutual information.

nbins
The number of bins passed to infotheo::discretize() if y is a matrix or casnet::ts_discrete()

doPlot
Produce a plot of the symbolic time series by calling plotRED_mif() (default = FALSE)

surTest
If TRUE, a surrogate will be conducted using simple surrogates. The surrogates will be created from the transition probabilities of the discretised time series, i.e. the probability of observing bin j when the current value is in bin j. The number of surrogates needed will be computed based on the value of the alpha parameter, conceived as a one-sided test: mi > 0.

alpha
The alpha level for the surrogate test (default = 0.05)

Value

The auto- or cross-mi function

See also

Examples

# Lags to evaluate mututal information
lags <- -10:30

# Auto-mutual information
y1 <- sin(seq(0, 100, by = 1/8)*pi)

(mif(data.frame(y1),lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.653815 1.927735 1.646619 1.573201 1.567816 1.678095 1.455365 1.497499 
#>       -2       -1        0        1        2        3        4        5 
#> 1.476553 2.868931 2.868931 2.868931 1.476553 1.497499 1.455365 1.678095 
#>        6        7        8        9       10       11       12       13 
#> 1.567816 1.573201 1.646619 1.927735 1.653815 1.550633 1.923477 1.562827 
#>       14       15       16       17       18       19       20       21 
#> 1.607086 1.621646 1.558334 1.859470 1.473869 1.623019 1.476240 1.515003 
#>       22       23       24       25       26       27       28       29 
#> 1.453358 1.751048 1.566279 1.964346 1.604698 1.668617 1.568317 1.564823 
#>       30 
#> 1.734763 
#> attr(,"miType")
#> [1] "I(X;X)"

# Cross-mututal information, y2 is a lagged version y1
y2 <- y1[10:801]

y <- data.frame(ts_trimfill(y1, y2, action = "trim.cut"))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.623019 1.473869 1.859470 1.558334 1.621646 1.607086 1.562827 1.923477 
#>       -2       -1        0        1        2        3        4        5 
#> 1.550633 1.653815 1.653815 1.653815 1.921974 1.687149 1.582642 1.610835 
#>        6        7        8        9       10       11       12       13 
#> 1.694766 1.456779 1.479514 1.468117 2.868171 1.483081 1.499478 1.471427 
#>       14       15       16       17       18       19       20       21 
#> 1.605462 1.578427 1.572016 1.660324 1.909035 1.644244 1.560248 1.923019 
#>       22       23       24       25       26       27       28       29 
#> 1.566571 1.604723 1.613996 1.696857 1.861291 1.474971 1.650919 1.465343 
#>       30 
#> 1.510893 
#> attr(,"miType")
#> [1] "I(X;Y)"

# Conditional mutual information, add some noise to y2 and add it as a 3rd column
y$s <- y2+rnorm(NROW(y2))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 1.862221 1.769316 1.996348 1.851036 1.995601 2.010456 1.962493 2.063345 
#>       -2       -1        0        1        2        3        4        5 
#> 1.884125 1.909442 1.909442 1.909442 2.041484 2.002678 1.979676 2.011610 
#>        6        7        8        9       10       11       12       13 
#> 2.025035 1.806844 1.774623 1.733806 2.516726 1.831044 1.879487 1.883610 
#>       14       15       16       17       18       19       20       21 
#> 1.932340 1.921755 1.880654 1.870835 2.012621 1.959015 1.997225 2.133363 
#>       22       23       24       25       26       27       28       29 
#> 1.988501 1.952195 1.882010 1.933072 1.973111 1.817211 1.944932 1.871809 
#>       30 
#> 1.896970 
#> attr(,"miType")
#> [1] "I(X;Y|Z)"

# Multi-information, the information of the entire multivariate series at each lag
y$y3 <- cumsum(rnorm(NROW(y)))
(mif(y,lags = lags))

#>      -10       -9       -8       -7       -6       -5       -4       -3 
#> 4.991668 4.986394 4.987553 4.989480 5.137644 5.137500 5.136654 5.135741 
#>       -2       -1        0        1        2        3        4        5 
#> 5.133076 5.126595 5.126595 5.126595 5.131354 5.135741 5.138442 5.141291 
#>        6        7        8        9       10       11       12       13 
#> 5.139578 4.991354 4.989461 4.988782 4.993674 4.996754 4.996337 4.995210 
#>       14       15       16       17       18       19       20       21 
#> 5.001949 5.142203 5.135363 5.138499 5.139541 5.138575 5.145333 5.149938 
#>       22       23       24       25       26       27       28       29 
#> 5.152429 5.151294 5.160289 5.159320 5.009102 5.008813 5.013780 5.020129 
#>       30 
#> 5.019313 
#> attr(,"miType")
#> [1] "I(X;Y;Z;...;N)"