Multiplex Recurrence Network

This function will create a Multiplex Recurrence Network from a list of igraph objects that can be considered the layers of a network. The layers must have the same number of nodes. There are two modes of operation: Layer similarity (MRNweightedBy is set to "InterLayerMI", "InterLayerCor", or "EdgeOvelap") and Layer importance (not implemented yet). The former generates weighted MRN based on Interlayer Mutual Information, Interlayer Correlation, or Edge Overlap.

mrn(
  layers,
  MRNweightedBy = c("InterlayerMI", "InterlayerCor", "Edgeoverlap")[1],
  win = NA,
  step = NA,
  overlap = NA,
  alignment = "r",
  cumulative = FALSE,
  doPlot = FALSE,
  silent = TRUE
)

Arguments

layers
A list of igraph objects representing the layers of the multiplex network. The layer networks must all have the same number of vertices.

MRNweightedBy
The measure to be used to evaluate the average structural similarities between the layers of the network. Valid options are: "InterLayerMI" (Mutual information based on similarity of the vertex degree across layers), "EdgeOverlap" (proportion of vertices sharing the same edges across layers). Choosing "InterLayerMI", "InterlayerCor", or "EdgeOverlap" will decide which measure is displayed in the plot of the Multiplex RN, all measures will always be returned in the numerical output.

win
The window size passed to ts_windower() in which to evaluate "InterLayerMI", "InterLayerCor", or "EdgeOvelap". (default = NA).

step
The stepsize for the sliding window (default = NA).

overlap
The window overlap passed to ts_windower() if MRNweightedBy is "InterLayerMI" or "EdgeOvelap". The value of step will be ignored if overlap is not NA. (default = NA).

alignment
Whether to right ("r"), center ("c"), or left ("l") align the window.

cumulative
To make the network represent cumulative time, set directed = TRUE and cumulative = TRUE. This will set the upper triangle of the recurrence matrix to 0 and ensures that the network edges represent recurrenct values that have occurred in the past relative to the current opbserved value. If directed = FALSE the argument is ignored (default = TRUE).

doPlot
Plot the multiplex recurrence network (default = FALSE).

silent
Silent-ish mode. (default = FALSE).

Value

A list object with fields:

interlayerMI - One or more matrices with edge weights between layers that represent the interlayer Mutual Information.

interlayerCor - One or more matrices with edge weights between layers that represent the Pearson correlation between vertex degrees of layers.

edgeOverlap - One or more matrices with edge weights between layers that represent the overlapping edges between layers.

meanValues - One or more matrices that represent the means and SDs of the interlayer Mutual Information, absolute interlayer correlation and edge overlap. Ther measure eo_joint refers to the number of edges shared among all layers of the MRN.

Examples

# Create some layers
library(igraph)

layers <- list(g1 = igraph::sample_smallworld(1, 100, 5, 0.05),
g2 = igraph::sample_smallworld(1, 100, 5, 0.5),
g3 = igraph::sample_smallworld(1, 100, 5, 1))

mrn(layers = layers)

#> 
#> ~~~o~~o~~casnet~~o~~o~~~
#> 
#> Welcome to the multiplex... in layer similarity mode!
#> 
#> 
#> ~~~o~~o~~casnet~~o~~o~~~

#> $interlayerMI
#> $interlayerMI$`window: 1 | start: 1 | stop: 100`
#>    g1       g2        g3
#> g1 NA 1.016009 0.9430457
#> g2 NA       NA 1.5627302
#> g3 NA       NA        NA
#> 
#> 
#> $edgeOverlap
#> $edgeOverlap$`window: 1 | start: 1 | stop: 100`
#>    g1    g2    g3
#> g1 NA 0.142 0.060
#> g2 NA    NA 0.049
#> g3 NA    NA    NA
#> 
#> 
#> $meanValues
#> $meanValues$`window: 1 | start: 1 | stop: 100`
#>    mi_mean    mi_sd    eo_mean      eo_sd eo_sum eo_all   eo_joint
#> 1 1.173928 0.338683 0.08366667 0.05081666   1500     16 0.01066667
#> 
#> 
#> $Nsize
#> [1] 100
#>