Multiplex Recurrence Network

This function will create a Multiplex Recurrence Network from a list of igraph objects that can be considered the layers of a network. The layers must have the same number of nodes. There are two modes of operation: Layer similarity (weightedBy is set to "InterLayerMI", "InterLayerCor", or "EdgeOvelap") and Layer importance (weightedBy is "AnisotropicCRQA"). The former generates weighted MRN based on Interlayer Mutual Information, Interlayer Correlation, or Edge Overlap, the latter examines the relative importance of each layer by assigning a rank to each vertex (time point), based on a vertex measure passed in argument MRNrankedBy.

mrn(
  layers,
  weightedBy = c("InterlayerMI", "InterlayerCor", "Edgeoverlap", "AnisotropicCRQA")[1],
  windowedWeights = c("none", "local", "cumulative")[1],
  CRQA_vertexSequence = "degree",
  CRQA_measure = c("RR", "DET", "LAM")[1],
  win = NA,
  step = NA,
  overlap = NA,
  alignment = "r",
  cumulative = TRUE,
  doPlot = FALSE,
  silent = TRUE
)

Arguments

layers
A list of igraph objects representing the layers of the multiplex network. The layer networks must all have the same number of vertices.

weightedBy
The measure to be used to evaluate the average structural similarities between the layers of the network. Valid options are: "InterLayerMI" (Mutual information based on similarity of the vertex degree across layers), "EdgeOverlap" (proportion of vertices sharing the same edges across layers), "RankorderDC" (Dynamic Complexity of the inter layer vertex rank order based on the vertex property/measure in AnisotropicCRQA). Choosing "InterLayerMI", "InterlayerCor", or "EdgeOverlap" will decide which measure is displayed in the plot of the Multiplex RN, all measures will always be returned in the numerical output.

windowedWeights
If a windowed analysis is conducted and the edges of the graphs in layers have a weight property, there are a number of different ways to handle the weights depending on the value of windowedWeights: "none", "local", and "cumulative". Value "none" will ignore the weights, "local" will limit the range of edges to those edges connecting the vertices contained within the window, "cumulative" will consider all edges connecting to vertices in the current window, including edges from vertices connecting from previous windows. (default = none)

CRQA_vertexSequence
If weightedBy = "AnisotropicCRQA", then CRQA_vertexSequence must be a valid igraph command that returns vertex properties, for example: "degree", "strength","hub_score","centr_degree", "transitivity", "betweenness". The appropriate measure type, e.g. for "directed", or "weighted" graphs, will be inferred from the graph properties of the 1st graph object in the layers list. For best results with weighted measures, assign a value to E(g)$weight for each g in layers. (default = "degree")

CRQA_measure
Which CRQA measures should be used c("RR", "DET", "LAM"),

win
The window size passed to ts_windower() in which to evaluate "InterLayerMI" or "EdgeOvelap". If weightedBy = "CRQA_vertexSequence", it will be the size of the right aligned window in which Dynamic Complexity will be computed using dc_win() (default = NA)

step
The stepsize for the sliding window (default = NA)

overlap
The window overlap passed to ts_windower() if weightedBy is "InterLayerMI" or "EdgeOvelap". The value of step will be ignored if overlap is not NA. (default = NA).

alignment
Whether to right ("r"), center ("c"), or left ("l") align the window.

cumulative
To make the network represent cumulative time, set directed = TRUE and cumulative = TRUE. This will set the upper triangle of the recurrence matrix to 0 and ensures that the network edges represent recurrenct values that have occurred in the past relative to the current opbserved value. If directed = FALSE the argument is ignored (default = TRUE).

doPlot
Plot the multiplex recurrence network (default = TRUE).

silent
Silent-ish mode

Value

A list object with fields:

interlayerMI - One or more matrices with edge weights between layers that represent the interlayer Mutual Information.

interlayerCor - One or more matrices with edge weights between layers that represent the Pearson correlation between vertex degrees of layers.

edgeOverlap - One or more matrices with edge weights between layers that represent the overlapping edges between layers.

meanValues - One or more matrices that represent the means and SDs of the interlayer Mutual Information, absolute interlayer correlation and edge overlap. Ther measure eo_joint refers to the number of edges shared among all layers of the MRN.