Plot Complexity Resonance Diagram

Plot Complexity Resonance Diagram

plotDC_res(df_win, win, useVarNames = TRUE, colOrder = TRUE,
  useTimeVector = NA, timeStamp = "01-01-1999", doPlot = TRUE,
  title = "Complexity Resonance Diagram", subtitle = "",
  xlabel = "Time", ylabel = "")

Arguments

df_win
A data frame containing series of Dynamic Complexity values obtained by running function dc_win()

useVarNames
Use the column names of df as variable names in the Complexity Resonance Diagram (default = TRUE)

colOrder
If TRUE, the order of the columns in df determines the of variables on the y-axis. Use FALSE for alphabetic/numeric order. Use NA to sort by by mean value of Dynamic Complexity (default = TRUE)

useTimeVector
Parameter used for plotting. A vector of length NROW(df), containing date/time information (default = NA)

timeStamp
If useTimeVector is not NA, a character string that can be passed to lubridate::stamp() to format the the dates/times passed in useTimeVector (default = "01-01-1999")

doPlot
If TRUE shows a Complexity Resonance Diagram of the Dynamic Complexity and returns an invisible ggplot2::ggplot() object. (default = FALSE)

title
A title for the plot.

subtitle
A subtitle for the plot.

xlabel
A label for the x-axis.

ylabel
A label for the y-axis.

Value

An invisible ggplot2 object.

See also