Create a Recurrence Network Matrix

This function serves as a wrapper for function rp(), it will add some attributes to the matrix related to network representation. These attributes will be used to decide which network type to generate (e.g. undirected, directed, weighted, etc.)

rn(
  y1,
  y2 = NULL,
  emDim = 1,
  emLag = 1,
  emRad = NULL,
  directed = FALSE,
  cumulative = TRUE,
  weighted = FALSE,
  weightedBy = c("si", "rt", "rf")[1],
  rescaleWeights = FALSE,
  fs = NA,
  includeDiagonal = FALSE,
  to.ts = NULL,
  order.by = NULL,
  to.sparse = FALSE,
  method = "Euclidean",
  targetValue = 0.05,
  returnGraph = FALSE,
  doPlot = FALSE,
  doEmbed = TRUE,
  silent = TRUE,
  ...
)

Arguments

y1
A numeric vector or time series

y2
A numeric vector or time series for cross recurrence

emDim
The embedding dimensions

emLag
The embedding lag

emRad
The threshold (emRad) to apply to the distance matrix to create a binary or weighted matrix. If NULL, an unthresholded matrix will be created (default = NULL)

directed
Should the matrix be considered to represent a directed network? (default = FALSE)

cumulative
To make the network represent cumulative time, set directed = TRUE and cumulative = TRUE. This will set the upper triangle of the recurrence matrix to 0 and ensures that the network edges represent recurrenct values that have occurred in the past relative to the current opbserved value. If directed = FALSE the argument is ignored (default = TRUE).

weighted
Should the matrix be considered to represent a weighted network? (default = FALSE)

weightedBy
After setting values smaller than emRad to 0, what should the recurrent values represent? The default is to use the state space similarity (distance/proximity) values as weights ("si"). Other option are "rt" for recurrence time and "rf" for recurrence time frequency, Because vertices represent time points in \(\epsilon\)-thresholded recurrence networks, a difference of two vertex-indices represents duration. If an edge e1 connects v1 and v10 then the recurrence time will be the difference of the vertex indices, 9, and the recurrence time frequency will be 1/9.

rescaleWeights
If set to TRUE and weighted = TRUE, all weight values will be rescaled to [0,1], where 0 means no recurrence relation and 1 the maximum weight value.

fs
Sample frequency: A numeric value interpreted as the number of observed samples per unit of time. If the weights represent recurrence times ("rt"), they will be divided by the value in fs. If the weights represent recurrence time frequencies ("rf"), they will be multiplied by the value of fs (default = NA)

includeDiagonal
Should the diagonal of the matrix be included when creating the network (default = FALSE)

to.ts
Should y1 and y2 be converted to time series objects?

order.by
If to.ts = TRUE, pass a vector of the same length as y1 and y2. It will be used as the time index, if NA the vector indices will be used to represent time.

to.sparse
Should sparse matrices be used?

method
Distance measure to use. Any option that is valid for argument method of proxy::dist(). Type proxy::pr_DB$get_entries() to se a list of all the options. Common methods are: "Euclidean", "Manhattan", "Minkowski", "Chebysev" (or the same but shorter: "L2","L1","Lp" and "max" distance) (default = "Euclidean")

targetValue
A value passed to est_radius(...,type="fixed", targetMeasure="RR") if is.na(emRad)==TRUE.

returnGraph
Return an igraph::igraph() object (default = FALSE)

doPlot
Plot the matrix by calling rp_plot() with default settings

doEmbed
If FALSE, a distance matrix will be returned that is not embedded by emDim and emLag. If y1 and/or y2 are data frames, the columns will be used as dimensions (default = TRUE)

silent
Silent-ish mode

...
Any paramters to pass to rn_plot() if doPlot = TRUE

Value

A (Coss-) Recurrence matrix that can be interpreted as an adjacency (or incidence) matrix.

See also