Recurrence Time Spectrum

Get the recurrence time distribution from a recurrence network.

rn_recSpec(
  RN,
  fitRange = NULL,
  fs = 1,
  doPlot = TRUE,
  returnPlot = FALSE,
  returnPLAW = FALSE,
  returnInfo = FALSE,
  silent = TRUE,
  noTitle = FALSE,
  tsName = "y"
)

Arguments

RN

A thresholded recurrence matrix generated by function rn()

fitRange

If NULL the entire range will be used for log-log slope. If a 2-element vector of integers, this will represent the range of recurrence times to use for fitting the log=log slope (e.g. c(1,50) would fit the first 50 recurrence times).

fs

Sample rate (default = 1)

doPlot

Should a plot of the recurrence time spectrum be produced?

returnPlot

Return ggplot2 object (default = FALSE)

returnPLAW

Return the power law data (default = FALSE)

returnInfo

Return all the data used in SDA (default = FALSE)

silent

Silent-ish mode

noTitle

Do not generate a title (only the subtitle)

tsName

Name of y added as a subtitle to the plot

Value

A vector of frequencies of recurrence times and a plot (if requested)

See also