Create a Distance Matrix

Create a Distance Matrix

rp(y1, y2 = NULL, emDim = 1, emLag = 1, emRad = NULL,
  to.ts = NULL, order.by = NULL, to.sparse = FALSE,
  weighted = FALSE, method = "Euclidean", doPlot = FALSE,
  silent = TRUE, ...)

Arguments

y1
A numeric vector or time series

y2
A numeric vector or time series for cross recurrence

emDim
The embedding dimensions

emLag
The embedding lag

emRad
The threshold (emRad) to apply to the distance matrix to create a binary or weighted matrix. If NULL, an unthresholded matrix will be created (default = NULL)

to.ts
Should y1 and y2 be converted to time series objects?

order.by
If to.ts = TRUE, pass a vector of the same length as y1 and y2. It will be used as the time index, if NA the vector indices will be used to represent time.

to.sparse
Should sparse matrices be used?

weighted
If FALSE a binary matrix will be returned. If TRUE every value larger than emRad will be 0, but values smaller than emRad will be retained (default = FALSE)

method
Distance measure to use. Any option that is valid for argument method of dist. Type proxy::pr_DB$get_entries() to se a list of all the options. Common methods are: "Euclidean", "Manhattan", "Minkowski", "Chebysev" (or the same but shorter: "L2","L1","Lp" and "max" distance) (default = "Euclidean")

doPlot
Plot the matrix by calling rp_plot with defult settings

silent
Silent-ish mode

...
Any paramters to pass to rp_plot if doPlot = TRUE

Value

A (Coss-) Recurrence matrix with attributes:

emdims1 and emdims2 - A matrix of surrogate dimensions

emdims1.name and emdims2.name - Names of surrogate dimensions

method and call - The distance method used by dist

weighetd - Whether a weighted matrix is returned

emDim, emLag and emRad - The embedding parameters

AUTO - Whether the matrix represents AUTO recurrence

See also