Diagonal Recurrence Profile

Diagonal Recurrence Profile

rp_diagProfile(
  RM,
  diagWin = NULL,
  xname = "X-axis",
  yname = "Y-axis",
  theiler = 0,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  DLmax = length(Matrix::diag(RM)) - 1,
  VLmax = length(Matrix::diag(RM)) - 1,
  HLmax = length(Matrix::diag(RM)) - 1,
  doShuffle = FALSE,
  y1 = NA,
  y2 = NA,
  Nshuffle = 19,
  AUTO = NULL,
  chromatic = FALSE,
  matrices = FALSE,
  doPlot = TRUE
)

Arguments

RM
A binary recurrence matrix

diagWin
Window around the line of synchrony

xname
Label for x-axis

yname
Label for y-axis

theiler
Use a theiler window around the line of identity / synchronisation to remove high auto-correlation at short time-lags (default = 0)

DLmin
Minimal diagonal line length (default = 2)

VLmin
Minimal vertical line length (default = 2)

HLmin
Minimal horizontal line length (default = 2)

DLmax
Maximal diagonal line length (default = length of diagonal -1)

VLmax
Maximal vertical line length (default = length of diagonal -1)

HLmax
Maximal horizontal line length (default = length of diagonal -1)

doShuffle
Should a shuffled baseline be calculated (default = FALSE)

y1
The original y1 time series

y2
The original y2 time series

Nshuffle
How many shuffled versions to make up the baseline? The default is 19, which is the minimum for a one-sided surrogate test.

AUTO
Auto-recurrence? (default = FALSE)

chromatic
Force chromatic RQA? (default = FALSE)

matrices
Return matrices? (default = FALSE)

doPlot
Plot

Value

A plot and/or the data for the plot