Line length distributions

Extract lengths of diagonal, vertical and horizontal line segments from a recurrence matrix.

rp_lineDist(
  RM,
  DLmin = 2,
  VLmin = 2,
  HLmin = 2,
  DLmax = length(Matrix::diag(RM)) - 1,
  VLmax = length(Matrix::diag(RM)) - 1,
  HLmax = length(Matrix::diag(RM)) - 1,
  d = NULL,
  theiler = NULL,
  invert = FALSE,
  AUTO = NULL,
  chromatic = FALSE,
  matrices = FALSE
)

Arguments

RM
A thresholded recurrence matrix (binary: 0 - 1)

DLmin
Minimal diagonal line length (default = 2)

VLmin
Minimal vertical line length (default = 2)

HLmin
Minimal horizontal line length (default = 2)

DLmax
Maximal diagonal line length (default = length of diagonal -1)

VLmax
Maximal vertical line length (default = length of diagonal -1)

HLmax
Maximal horizontal line length (default = length of diagonal -1)

d
Vector of diagonals to be extracted from matrix RP before line length distributions are calculated. A one element vector will be interpreted as a windowsize, e.g., d = 50 will extract the diagonal band -50:50. A two element vector will be interpreted as a band, e.g. d = c(-50,100) will extract diagonals -50:100. If length(d) > 2, the numbers will be interpreted to refer to individual diagonals, d = c(-50,50,100) will extract diagonals -50,50,100.

theiler
Size of the theiler window, e.g. theiler = 1 removes diagonal bands -1,0,1 from the matrix. If length(d) is NULL, 1 or 2, the theiler window is applied before diagonals are extracted. The theiler window is ignored if length(d)>2, or if it is larger than the matrix or band indicated by parameter d.

invert
Relevant for Recurrence Time analysis: Return the distribution of 0 valued segments in nonzero diagonals/verticals/horizontals. This indicates the time between subsequent line structures.

AUTO
Is this an AUTO RQA?

chromatic
Chromatic RQA?

matrices
Return the matrices?

Value

A list object with distributions of line lengths. If matrices = TRUE datafr are returned whose columns represent the nonzero diagonals, verticals, or, horizontals.

Details

Based on the Matlab function linedists by Stefan Schinkel, Copyright (C) 2009 Stefan Schinkel, University of Potsdam, http://www.agnld.uni-potsdam.de

References: S. Schinkel, N. Marwan, O. Dimigen & J. Kurths (2009): "Confidence Bounds of recurrence-based complexity measures Physics Letters A, 373(26), pp. 2245-2250

Copyright (C) 2009 Stefan Schinkel, University of Potsdam http://www.agnld.uni-potsdam.de

See also